식단 관리 관련 모든 페이지의 데이터를 정수화

+++ ManagingDiet/src/main/java/com/diet/biz/dietProgram/DietProgramService.java
@@ -7,7 +7,7 @@ import com.diet.biz.user.UserVO;
 public interface DietProgramService {
 	
 	// 사용자 스펙을 기반으로 한 기초대사량 계산
-	double dietStep1(KcalVO vo);
+	int dietStep1(KcalVO vo);
 	
 	// 기초대사량에 근거한 칼로리 계산
 	int dietStep2(KcalVO vo);
